refactor(mind-map): extract initEvent function
